requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
aiohttp==3.9.1
pyyaml==6.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
fpdf==1.7.2
loguru==0.7.2
typer==0.9.0
pytest==7.4.4
pytest-asyncio==0.23.3
